Make all commands work with both v1 and v2 manifests
This wires the cli to the new installer code for both v1 and v2
distribution manifests. v1 installations are upgraded to v2
automatically.

I copied test cases from multirust.sh into tests/cli-v1.rs,
tests/cli-v2.rs and tests/cli-misc.rs. Many of them unrelated to
getting installation from the distribution server working are still
unported and marked #[ignore]. There is a lot of setup necessary for
these tests and they are quite slow. I've been testing with --release.

The metadata format is sloppy right now.

* rustlib/rust-installer-version - Still the same as rust-installer
* rustlib/components - Components has its own component list that is
  the same as rust-installer, and I think needed for v1 manifest
  compatibility
* rustlib/manifest-$component - The component manifest, per rust-installer
* rustlib/multirust-config.toml - The multirust distribution
  configuration, which currently contains its own component/target
  list per the v2 manifests, and is versioned.
* rustlib/multirust-channel-manifest.toml - A stored copy of the
  distribution manifest.

+++ rust-install/src/utils/raw.rs
@@ -199,6 +199,13 @@ pub fn download_file<P: AsRef<Path>>(url: hyper::Url,
     // The file scheme is mostly for use by tests to mock the dist server
     if url.scheme == "file" {
         let src = try!(url.to_file_path().map_err(|_| DownloadError::FilePathParse));
+        if !is_file(&src) {
+            // Because some of multirust's logic depends on checking
+            // the error when a downloaded file doesn't exist, make
+            // the file case return the same error value as the
+            // network case.
+            return Err(DownloadError::Status(hyper::status::StatusCode::NotFound));
+        }
         try!(fs::copy(&src, path.as_ref()).map_err(|e| DownloadError::File(e)));
 
         if let Some(ref mut h) = hasher {
